+++ http_check/tests/test_http_integration.py
@@ -2,6 +2,7 @@
 # All rights reserved
 # Licensed under a 3-clause BSD style license (see LICENSE)
 import os
+import re
 import sys
 from collections import OrderedDict
 
@@ -174,7 +175,7 @@ def test_check_cert_expiration_self_signed(http_check):
     if PY2:
         assert "certificate verify failed" in msg
     else:
-        assert "certificate verify failed: self signed certificate" in msg
+        assert re.search("certificate verify failed: self[- ]signed certificate", msg)
 
 
 @pytest.mark.usefixtures("dd_environment")
@@ -539,3 +540,41 @@ def test_tls_config_ok(dd_run_check, instance, check_hostname):
     )
     tls_context = check.get_tls_context()
     assert tls_context.check_hostname is check_hostname
+
+
+@pytest.mark.parametrize(
+    'headers',
+    [
+        pytest.param({'content-type': 'application/json'}),
+        pytest.param({'Content-Type': 'application/json'}),
+        pytest.param({'CONTENT-TYPE': 'text/html'}),
+        pytest.param({}),
+    ],
+)
+def test_case_insensitive_header_content_type(dd_run_check, headers):
+    """
+    Test that `Content-Type` is accessible from the headers dict regardless of letter case.
+    We're only testing `Content-Type` for a non-GET method because that's the only header field
+    that applies a default header value if omitted.
+    """
+    instance = {
+        'name': 'foobar',
+        'url': 'http://something.com',
+        'method': 'POST',
+        'headers': headers,
+        'data': {'foo': 'bar'},
+    }
+    default_headers = {
+        'User-Agent': 'Datadog Agent/0.0.0',
+        'Accept': '*/*',
+        'Accept-Encoding': 'gzip, deflate',
+        'Content-Type': 'application/x-www-form-urlencoded',
+    }
+    check = HTTPCheck('http_check', {'ca_certs': 'foo'}, [instance])
+
+    dd_run_check(check)
+
+    if headers == {}:
+        assert check.http.options["headers"] == default_headers
+    else:
+        assert check.http.options["headers"] == headers