+++ sqlserver/CHANGELOG.md
@@ -1,5 +1,24 @@
 # CHANGELOG - sqlserver
 
+## 11.0.2 / 2023-03-21
+
+* [Fixed] Increase command_timeout default from 5s to 10s to prevent timeouts on instances with large SQL caches. See [#14195](https://github.com/DataDog/integrations-core/pull/14195).
+
+## 11.0.1 / 2023-03-08
+
+* [Fixed] Do not double emit instance metrics when autodiscovery is enabled. See [#14115](https://github.com/DataDog/integrations-core/pull/14115).
+
+## 11.0.0 / 2023-03-03
+
+* [Added] Collect program_name for SQL Server activity. See [#13953](https://github.com/DataDog/integrations-core/pull/13953).
+* [Fixed] sqlserver.database.state not sent for all databases. See [#13735](https://github.com/DataDog/integrations-core/pull/13735).
+* [Fixed] Bug Fix: Azure SQL DB database name tags properly applied on perf metrics. See [#13757](https://github.com/DataDog/integrations-core/pull/13757).
+* [Changed] Lower the frequency of query metrics collection. See [#14033](https://github.com/DataDog/integrations-core/pull/14033).
+
+## 10.1.3 / 2023-02-01 / Agent 7.43.0
+
+* [Fixed] Fix sqlserver.database.state not being sent for all databases. See [#13735](https://github.com/DataDog/integrations-core/pull/13735).
+
 ## 10.1.2 / 2023-01-26
 
 * [Fixed] Bug Fix: Azure SQL DB database name tags properly applied on perf metrics. See [#13757](https://github.com/DataDog/integrations-core/pull/13757).