+++ zk/CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG - zk
 
+## 4.2.2 / 2023-02-13
+
+* [Fixed] Ensure Zookeeper command execution in proxy environments. See [#13935](https://github.com/DataDog/integrations-core/pull/13935).
+
 ## 4.2.1 / 2022-12-09 / Agent 7.42.0
 
 * [Fixed] Stop using deprecated `distutils.version` classes. See [#13408](https://github.com/DataDog/integrations-core/pull/13408).